refactor: separate platform ID from messages
