+++ src/popup/presets.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import type { PlatformId } from '../messages';
+import type { PlatformId } from '../types/platform';
 import { DEFAULT_SELECTED_PLATFORMS } from './platforms';
 import {
   applyPresetSelection,
