+++ src/types/platform.ts
@@ -1,0 +1,13 @@
+/** Tutti が投稿・診断・履歴で識別する platform ID。 */
+export type PlatformId =
+  | 'x'
+  | 'bluesky'
+  | 'threads'
+  | 'mastodon'
+  | 'misskey'
+  | 'tumblr'
+  | 'pixiv'
+  | 'deviantart'
+  | 'instagram'
+  | 'tiktok'
+  | 'youtube';
