+++ src/utils/reply-compose.ts
@@ -1,4 +1,4 @@
-import type { PlatformId } from '../messages';
+import type { PlatformId } from '../types/platform';
 import {
   elementTextMatches,
   isElementDisabled,
