+++ src/utils/user-detect.ts
@@ -1,4 +1,4 @@
-import type { PlatformId } from '../messages';
+import type { PlatformId } from '../types/platform';
 import { sleep } from './dom';
 
 /**
